api/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.28
python-dotenv==1.0.1
